Revert "move query params handling into useSettings"
This reverts commit 60cf719a4322113dfc06fd5c043d19fb71c8ff5a.
