+++ webui/react/src/hooks/useSettings.ts
@@ -1,6 +1,6 @@
 import * as t from 'io-ts';
 import { useObservable } from 'micro-observables';
-import { useCallback, useContext, useEffect, useLayoutEffect, useMemo, useRef } from 'react';
+import { useCallback, useContext, useEffect, useLayoutEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { updateUserSetting } from 'services/api';
@@ -162,17 +162,12 @@ const queryToSettings = <T>(config: SettingsConfig<T>, query: string) => {
 };
 
 const useSettings = <T>(config: SettingsConfig<T>): UseSettingsReturn<T> => {
-  const { isLoading: isLoadingOb, state: stateOb } = useContext(UserSettings);
-
-  const querySettings = useRef('');
-  useEffect(() => {
-    const url = window.location.search.substring(/^\?/.test(location.search) ? 1 : 0);
-    querySettings.current = url;
-  }, []);
-  const clearQuerySettings = useCallback(() => {
-    querySettings.current = '';
-  }, []);
-
+  const {
+    isLoading: isLoadingOb,
+    querySettings,
+    state: stateOb,
+    clearQuerySettings,
+  } = useContext(UserSettings);
   const initialLoading = useObservable(isLoadingOb);
   const derivedOb = useMemo(
     () => stateOb.select((s) => s.get(config.storagePath)),
@@ -310,15 +305,15 @@ const useSettings = <T>(config: SettingsConfig<T>): UseSettingsReturn<T> => {
 
   // parse navigation url to state
   useEffect(() => {
-    if (!querySettings.current) return;
+    if (!querySettings) return;
 
-    const parsedSettings = queryToSettings<T>(config, querySettings.current);
+    const parsedSettings = queryToSettings<T>(config, querySettings);
     updateSettings(parsedSettings);
-  }, [config, updateSettings]);
+  }, [config, querySettings, updateSettings]);
 
   useLayoutEffect(() => {
     if (initialLoading) return;
-    clearQuerySettings();
+    setTimeout(clearQuerySettings, 200);
     return derivedOb.subscribe(async (cur, prev) => {
       if (!cur || !currentUser || isEqual(cur, prev)) return;
