+++ webui/react/src/hooks/useSettingsProvider.tsx
@@ -1,6 +1,6 @@
 import { Map } from 'immutable';
 import { observable, useObservable, WritableObservable } from 'micro-observables';
-import React, { createContext, useEffect, useState } from 'react';
+import React, { createContext, useCallback, useEffect, useRef, useState } from 'react';
 
 import Spinner from 'components/Spinner';
 import { getUserSetting } from 'services/api';
@@ -21,12 +21,16 @@ type UserSettingsState = Map<string, Settings>;
 export type Settings = { [key: string]: any }; //TODO: find a way to use a better type here
 
 type UserSettingsContext = {
+  clearQuerySettings: () => void;
   isLoading: WritableObservable<boolean>;
+  querySettings: string;
   state: WritableObservable<UserSettingsState>;
 };
 
 export const UserSettings = createContext<UserSettingsContext>({
+  clearQuerySettings: () => undefined,
   isLoading: observable(false),
+  querySettings: '',
   state: observable(Map<string, Settings>()),
 });
 
@@ -35,6 +39,7 @@ export const SettingsProvider: React.FC<React.PropsWithChildren> = ({ children }
   const isAuthChecked = useObservable(authStore.isChecked);
   const [canceler] = useState(new AbortController());
   const [isLoading] = useState(() => observable(true));
+  const querySettings = useRef('');
   const [settingsState] = useState(() => observable(Map<string, Settings>()));
 
   useEffect(() => {
@@ -75,13 +80,25 @@ export const SettingsProvider: React.FC<React.PropsWithChildren> = ({ children }
     return () => canceler.abort();
   }, [canceler, isAuthChecked, isLoading, settingsState]);
 
+  useEffect(() => {
+    const url = window.location.search.substring(/^\?/.test(location.search) ? 1 : 0);
+
+    querySettings.current = url;
+  }, []);
+
+  const clearQuerySettings = useCallback(() => {
+    querySettings.current = '';
+  }, []);
+
   return (
     <Spinner
       spinning={useObservable(isLoading) && !(isAuthChecked && !currentUser)}
       tip="Loading Page">
       <UserSettings.Provider
         value={{
+          clearQuerySettings,
           isLoading: isLoading,
+          querySettings: querySettings.current,
           state: settingsState,
         }}>
         {children}
